Add config.yml and pertinant logic
A lot of the config stuff is handled by spigot but ConfigManager is there because it allows me to interface with the config in the same way I do with BSIPA plugins. Thanks Kotlin for built-in getter and setter logic.

- Add rudimentary set command to change config properties
- Change GameManager and StartTimer to respect the config instead of various constants.
- Make a reference to GameManager in DeathShift.onEnable() to reduce cold start time for players.

Is any of this good practise? I have no idea.