Add game sounds
I'm pleased with the way I handled the playing of sounds in SoundUtil, very cool. I do not plan on adding any config options for the sounds since I don't want to add a bunch of conditions to evey call to SoundUtil, and I don't want SoundUtil itself to have conditions. I like things the way they are :D